+++ src/pages/Dashboard/Product.js
@@ -1,20 +1,22 @@
 import React from 'react';
 
-const Product = ({product, refetch}) => {
-    const {name, price, available_quantity, _id} = product ;
-    const deleteProduct = id => { 
-        fetch(`https://tranquil-tundra-16871.herokuapp.com/product/${id}`, {
-            method: "DELETE",
-            headers: {
-                authorization : `Bearer ${localStorage.getItem('accessToken')}`
-            }
-        })
-        .then(res => res.json())
-        .then(data => {
-            refetch();
-            console.log(data)
-        })
-    } 
+const Product = ({ product, refetch }) => {
+    const { name, price, available_quantity, _id } = product;
+    const deleteProduct = id => {
+        const sure = window.confirm('Are Your Sure');
+        if (sure) {
+            fetch(`https://tranquil-tundra-16871.herokuapp.com/product/${id}`, {
+                method: "DELETE",
+                headers: {
+                    authorization: `Bearer ${localStorage.getItem('accessToken')}`
+                }
+            })
+                .then(res => res.json())
+                .then(data => {
+                    refetch();
+                })
+        }
+    }
     return (
         <tr>
             <td>{name}</td>
